Fix call to ynh_script_progression in upgrade.j2
In one place in upgrade.j2, ynh_script_progression
is called with a --message flag instead of just passing the message as
an argument.
This results in "--message" being included in the output when running
the upgrade script:
`Info: [####++++............] > --message=Upgrading source files...`

I searched the repo for "--message" and found no other occurences.